+++ src/main-gui.cpp
@@ -45,10 +45,8 @@ int main(int argc, char *argv[]) {
   engine.load(QUrl(QStringLiteral("qrc:/qml/screens/main.qml")));
   if (engine.rootObjects().isEmpty()) return -1;
 
-  // Create Websocket server object.
+  // Create Websocket server object and connect close button signal
   qmlsystem.setWSServer();
-
-  // Create 
   QObject::connect(&app, SIGNAL(aboutToQuit()), &qmlsystem, SLOT(cleanAndClose()));
   return app.exec();
 }